Added a Road graphic, sorted out files and classes
