+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
     "name": "Timing Traffic",
-    "slug": "Traffic 'round the clock",
+    "slug": "TrafficTiming",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
