+++ app/assets/Road.js
@@ -1,0 +1,28 @@
+import React from "react";
+import Svg, { Path } from "react-native-svg";
+
+/**
+ * Made with Pixels to Svg https://codepen.io/shshaw/pen/XbxvNj
+ */
+export default () => {
+    return (
+        <Svg viewBox="0 0 32 64" shapeRendering="crispEdges">
+            <Path
+                stroke="#667384"
+                d="M4 0h24M2 1h2M28 1h2M1 2h1M30 2h1M1 3h1M30 3h1M0 4h1M9 4h14M31 4h1M0 5h1M7 5h2M22 5h3M31 5h1M0 6h1M24 6h2M31 6h1M0 7h1M25 7h1M31 7h1M0 8h1M26 8h1M31 8h1M0 9h1M26 9h1M31 9h1M0 10h1M31 10h1M0 11h1M31 11h1M0 12h1M31 12h1M0 13h1M31 13h1M0 14h1M31 14h1M0 15h1M31 15h1M0 16h1M31 16h1M0 17h1M31 17h1M0 18h1M31 18h1M0 19h1M31 19h1M0 20h1M31 20h1M0 21h1M31 21h1M0 22h1M31 22h1M0 23h1M31 23h1M0 24h1M31 24h1M0 25h1M31 25h1M0 26h1M31 26h1M0 27h1M31 27h1M0 28h1M31 28h1M0 29h1M31 29h1M0 30h1M31 30h1M0 31h1M31 31h1M0 32h1M31 32h1M0 33h1M31 33h1M0 34h1M31 34h1M0 35h1M31 35h1M0 36h1M31 36h1M0 37h1M31 37h1M0 38h1M31 38h1M0 39h1M31 39h1M0 40h1M31 40h1M0 41h1M31 41h1M0 42h1M31 42h1M0 43h1M31 43h1M0 44h1M31 44h1M0 45h1M31 45h1M0 46h1M31 46h1M0 47h1M31 47h1M0 48h1M31 48h1M0 49h1M31 49h1M0 50h1M31 50h1M0 51h1M31 51h1M0 52h1M31 52h1M0 53h1M31 53h1M0 54h1M5 54h1M31 54h1M0 55h1M5 55h1M31 55h1M0 56h1M6 56h1M31 56h1M0 57h1M7 57h1M31 57h1M0 58h1M8 58h2M22 58h2M31 58h1M0 59h1M10 59h12M31 59h1M1 60h1M30 60h1M1 61h1M30 61h1M2 62h2M28 62h2M4 63h24"
+            />
+            <Path
+                stroke="#848d9e"
+                d="M4 1h24M2 2h4M26 2h4M2 3h3M6 3h20M27 3h3M1 4h8M23 4h8M1 5h2M4 5h3M25 5h3M29 5h2M1 6h1M3 6h3M26 6h3M30 6h1M1 7h1M3 7h3M26 7h3M30 7h1M1 8h1M3 8h2M27 8h2M30 8h1M1 9h1M3 9h2M27 9h2M30 9h1M1 10h1M3 10h1M28 10h1M30 10h1M1 11h1M3 11h1M28 11h1M30 11h1M1 12h1M3 12h1M28 12h1M30 12h1M1 13h1M3 13h1M28 13h1M30 13h1M1 14h1M3 14h1M28 14h1M30 14h1M1 15h1M3 15h1M28 15h1M30 15h1M1 16h3M28 16h3M1 17h3M28 17h3M1 18h1M3 18h1M28 18h1M30 18h1M1 19h1M3 19h1M28 19h1M30 19h1M1 20h1M3 20h1M28 20h1M30 20h1M1 21h1M3 21h1M28 21h1M30 21h1M1 22h1M3 22h1M28 22h1M30 22h1M1 23h1M3 23h1M28 23h1M30 23h1M1 24h1M3 24h1M28 24h1M30 24h1M1 25h1M3 25h1M28 25h1M30 25h1M1 26h1M3 26h1M28 26h1M30 26h1M1 27h1M3 27h1M28 27h1M30 27h1M1 28h1M3 28h1M28 28h1M30 28h1M1 29h1M3 29h1M28 29h1M30 29h1M1 30h1M3 30h1M28 30h1M30 30h1M1 31h1M3 31h1M28 31h1M30 31h1M1 32h3M28 32h3M1 33h3M28 33h3M1 34h1M3 34h1M28 34h1M30 34h1M1 35h1M3 35h1M28 35h1M30 35h1M1 36h1M3 36h1M28 36h1M30 36h1M1 37h1M3 37h1M28 37h1M30 37h1M1 38h1M3 38h1M28 38h1M30 38h1M1 39h1M3 39h1M28 39h1M30 39h1M1 40h1M3 40h1M28 40h1M30 40h1M1 41h1M3 41h1M28 41h1M30 41h1M1 42h1M3 42h1M28 42h1M30 42h1M1 43h1M3 43h1M28 43h1M30 43h1M1 44h1M3 44h1M28 44h1M30 44h1M1 45h1M3 45h1M28 45h1M30 45h1M1 46h1M3 46h1M28 46h1M30 46h1M1 47h1M3 47h1M28 47h1M30 47h1M1 48h3M28 48h3M1 49h3M28 49h3M1 50h1M3 50h1M28 50h1M30 50h1M1 51h1M3 51h1M28 51h1M30 51h1M1 52h1M3 52h1M28 52h1M30 52h1M1 53h1M3 53h1M28 53h1M30 53h1M1 54h1M3 54h2M27 54h2M30 54h1M1 55h1M3 55h2M27 55h2M30 55h1M1 56h1M3 56h3M26 56h3M30 56h1M1 57h1M3 57h4M25 57h4M30 57h1M1 58h2M4 58h4M24 58h4M29 58h2M1 59h9M22 59h9M2 60h3M6 60h20M27 60h3M2 61h4M26 61h4M4 62h24"
+            />
+            <Path
+                stroke="#e2b58c"
+                d="M6 2h20M5 3h1M26 3h1M3 5h1M28 5h1M2 6h1M29 6h1M2 7h1M29 7h1M2 8h1M29 8h1M2 9h1M29 9h1M2 10h1M29 10h1M2 11h1M29 11h1M2 12h1M29 12h1M2 13h1M29 13h1M2 14h1M29 14h1M2 15h1M29 15h1M2 18h1M29 18h1M2 19h1M29 19h1M2 20h1M29 20h1M2 21h1M29 21h1M2 22h1M29 22h1M2 23h1M29 23h1M2 24h1M29 24h1M2 25h1M29 25h1M2 26h1M29 26h1M2 27h1M29 27h1M2 28h1M29 28h1M2 29h1M29 29h1M2 30h1M29 30h1M2 31h1M29 31h1M2 34h1M29 34h1M2 35h1M29 35h1M2 36h1M29 36h1M2 37h1M29 37h1M2 38h1M29 38h1M2 39h1M29 39h1M2 40h1M29 40h1M2 41h1M29 41h1M2 42h1M29 42h1M2 43h1M29 43h1M2 44h1M29 44h1M2 45h1M29 45h1M2 46h1M29 46h1M2 47h1M29 47h1M2 50h1M29 50h1M2 51h1M29 51h1M2 52h1M29 52h1M2 53h1M29 53h1M2 54h1M29 54h1M2 55h1M29 55h1M2 56h1M29 56h1M2 57h1M29 57h1M3 58h1M28 58h1M5 60h1M26 60h1M6 61h20"
+            />
+            <Path
+                stroke="#414852"
+                d="M6 6h1M6 7h1M5 8h1M5 9h1M4 10h1M27 10h1M4 11h1M27 11h1M4 12h1M27 12h1M4 13h1M27 13h1M4 14h1M27 14h1M4 15h1M27 15h1M4 16h1M27 16h1M4 17h1M27 17h1M4 18h1M27 18h1M4 19h1M27 19h1M4 20h1M27 20h1M4 21h1M27 21h1M4 22h1M27 22h1M4 23h1M27 23h1M4 24h1M27 24h1M4 25h1M27 25h1M4 26h1M27 26h1M4 27h1M27 27h1M4 28h1M27 28h1M4 29h1M27 29h1M4 30h1M27 30h1M4 31h1M27 31h1M4 32h1M27 32h1M4 33h1M27 33h1M4 34h1M27 34h1M4 35h1M27 35h1M4 36h1M27 36h1M4 37h1M27 37h1M4 38h1M27 38h1M4 39h1M27 39h1M4 40h1M27 40h1M4 41h1M27 41h1M4 42h1M27 42h1M4 43h1M27 43h1M4 44h1M27 44h1M4 45h1M27 45h1M4 46h1M27 46h1M4 47h1M27 47h1M4 48h1M27 48h1M4 49h1M27 49h1M4 50h1M27 50h1M4 51h1M27 51h1M4 52h1M27 52h1M4 53h1M27 53h1M26 54h1M26 55h1M25 56h1M24 57h1"
+            />
+        </Svg>
+    );
+};
