+++ app/components/index.js
@@ -1,0 +1,4 @@
+export { default as Timer } from "./Timer";
+export { default as ButtonsRow } from "./RoundButton";
+export { default as LapsTable } from "./LapsTable";
+export { default as RoundButton } from "./RoundButton";
