+++ package.json
@@ -9,6 +9,10 @@
     "web": "expo start --web",
     "eject": "expo eject"
   },
+  "resolutions": {
+    "graceful-fs": "4.2.4",
+    "react-error-overlay": "^6.0.9"
+  },
   "dependencies": {
     "@expo/webpack-config": "~0.16.2",
     "expo": "~45.0.0",
@@ -16,12 +20,19 @@
     "moment": "^2.29.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-error-overlay": "6.0.9",
     "react-native": "0.68.2",
+    "react-native-safe-area-context": "4.2.4",
+    "react-native-screens": "~3.11.1",
+    "react-native-svg": "^12.3.0",
     "react-native-web": "^0.17.7",
+    "react-navigation": "^4.4.4",
+    "react-scripts": "^5.0.1",
     "webpack-dev-server": "~3.11.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9"
+    "@babel/core": "^7.12.9",
+    "react-error-overlay": "^6.0.9"
   },
   "private": true
 }
